Splits tensor intiializer functions into separate files
